+++ internal/access/formatter/objects/you.go
@@ -2,13 +2,13 @@ package objects
 
 // You is a unit object. It is most often the win or lose condition in a game.
 type You struct {
-	Owner uint8
-	HP    uint8
-	State uint8
+	Owner int
+	HP    int
+	State int
 }
 
 // NewYou returns a new You object
-func NewYou(owner, hp, state uint8) *You {
+func NewYou(owner, hp, state int) *You {
 	return &You{
 		Owner: owner,
 		HP:    hp,
@@ -17,11 +17,11 @@ func NewYou(owner, hp, state uint8) *You {
 }
 
 // GetUnitType see function from Unit
-func (you *You) GetUnitType() uint8 {
+func (you *You) GetUnitType() int {
 	return UnitTypeYou
 }
 
 // GetUnitState see function from Unit
-func (you *You) GetUnitState() uint8 {
+func (you *You) GetUnitState() int {
 	return you.State
 }
